+++ nodejs/systemHandler.js
@@ -69,6 +69,10 @@ class SystemHandler {
                 return handleNTP.call(this);
             })
             .then(() => {
+                logger.fine('Checking ManagementRoute.');
+                return handleManagementRoute.call(this);
+            })
+            .then(() => {
                 logger.fine('Checking hostname.');
                 return handleHostname.call(this);
             })
@@ -381,6 +385,35 @@ function handleLicensePool(license) {
         .catch(err => Promise.reject(err));
 }
 
+function handleManagementRoute() {
+    const promises = [];
+    doUtil.forEach(this.declaration, 'ManagementRoute', (tenant, managementRoute) => {
+        if (managementRoute && managementRoute.name) {
+            const routeBody = {
+                name: managementRoute.name,
+                partition: tenant,
+                gw: managementRoute.gw,
+                network: managementRoute.network,
+                mtu: managementRoute.mtu
+            };
+
+            if (managementRoute.type) {
+                routeBody.type = managementRoute.type;
+            }
+
+            promises.push(
+                this.bigIp.createOrModify(PATHS.ManagementRoute, routeBody, null, cloudUtil.MEDIUM_RETRY)
+            );
+        }
+    });
+
+    return Promise.all(promises)
+        .catch((err) => {
+            logger.severe(`Error creating management routes: ${err.message}`);
+            throw err;
+        });
+}
+
 function createOrUpdateUser(username, data) {
     let userEndpoint = '/tm/auth/user';
     if (this.bigIp.isBigIq()) {
@@ -457,6 +490,13 @@ function disableDhcpOptions(optionsToDisable) {
                 }
             );
         })
+        .then(() => this.bigIp.list('/tm/sys/global-settings'))
+        .then((globalSettings) => {
+            // If DHCP is disabled on the device do NOT attempt to restart it
+            if (globalSettings && globalSettings.mgmtDhcp === 'disabled') {
+                requiresDhcpRestart = false;
+            }
+        })
         .then(() => {
             if (!requiresDhcpRestart) {
                 return Promise.resolve();
@@ -486,7 +526,7 @@ function restartDhcp() {
     )
         .then(() => {
             function isDhcpRunning() {
-                return this.bigIp.list('/tm/sys/service/dhclient/stats')
+                return this.bigIp.list('/tm/sys/service/dhclient/stats', undefined, cloudUtil.NO_RETRY)
                     .then((dhcpStats) => {
                         if (dhcpStats.apiRawValues
                             && dhcpStats.apiRawValues.apiAnonymous